requirements.txt
# COre Dependencies
pandas
numpy
scikit-learn

# versioning and experiment tracking
dvc
dvc-s3
dagshub
mlflow
python-dotenv
